deep_agent/requirements.txt
deepagents
langgraph
langchain
langchain-fireworks
langchain-openai
redis
tavily-python
python-dotenv
tavily-python
